\docType{data}
\name{envirX}
\alias{envirX}
\title{CNV simulated environmental variables}
\description{
  The simulated environmental file which contains the
  possible environmental variables. The order of the row in
  this file must consistent with the second column in FAM
  file.
}
\author{
  Meiling Liu
}
\keyword{datasets}

