\docType{data}
\name{fam}
\alias{fam}
\title{CNV simulated data}
\description{
  The simulated FAM file. The first six columns of FAM file
  are mandatory: Family ID, Individual ID, Paternal ID,
  Maternal ID, Sex (1=male; 2=female; other=unknown) and
  Phenotype.
}
\author{
  Meiling Liu
}
\keyword{datasets}

