\docType{package}
\name{PedCNV-package}
\alias{PedCNV-package}
\title{CNV association implementation}
\description{
  A package to perform robust quantitative traits
  association testing of copy number variants. It provides
  two methods for association study: first, the observed
  probe intensity measurement can be directly used to
  detect the association of CNV with phenotype of interest.
  Second, the most probable copy number is estimated with
  the proposed likelihood and the association of the most
  probable copy number with phenotype is tested. Also, it
  can be used to determine the optimal clustering number
  and clustering assignment for each individuals. This
  method can be applied to both the independent and
  correlated population.
}
\details{
  \tabular{ll}{ Package: \tab PedCNV\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-09-03\cr
  License: \tab MIT \cr Main functions: \tab AssoTestProc
  \cr \tab ClusProc \cr \tab STE \cr \tab STIM \cr \tab
  print.asso\cr \tab print.clus\cr \tab plot.clus\cr }
}
\author{
  Meiling Liu, Sungho Won and Weicheng Zhu
}
\references{
  On the association analysis of CNV data: fast and
  efficient method with family-based samples
}

