\name{plot.clust}
\alias{plot.clust}
\title{TODO}
\usage{
  \method{plot}{clust} (x,
    type = c("histo", "scat", "sil"), adjust = TRUE, ...)
}
\arguments{
  \item{x}{TODO}

  \item{type}{TODO}

  \item{adjust}{TODO}

  \item{...}{TODO}
}
\value{
  TODO
}
\description{
  TODO
}
\details{
  TODO
}
\author{
  Meiling
}

