\docType{data}
\name{simudat}
\alias{simudat}
\title{CNV simulated data}
\description{
  Simulated data. It contains CNV intensity measurements,
  environment variables and FAM file which follows the same
  format defined in PLINK.
}
\examples{
data(simudat)
}
\author{
  Meiling Liu
}
\keyword{datasets}

