\docType{data}
\name{phi}
\alias{phi}
\title{Empirical correlation matrix}
\description{
  Empirical/kinship correlation matrix between individuals.
  This correlation matrix can be calculated based on the
  familial relationship between individuals or large-scale
  SNP data by omic data analysis toolkit FQLS. The free
  software FQLS can be downloaded from
  \url{http://biostat.cau.ac.kr/fqls/}. If correlation
  matrix is estimated with the large-scale SNP data, the
  proposed method becomes robust under the presence of
  population substructure.
}
\examples{
data(phi)
}
\author{
  Meiling Liu
}
\references{
  FQLS \url{http://biostat.cau.ac.kr/fqls/}
}
\keyword{datasets}

